Set up new project structure
The application needed to be updated to python3, so I'm taking this
chance to also update the project structure.  The preferred method for
running Flask applications now uses app factories, which allows for
production deployment.  Using app.run() is not preferred, since that
uses the built-in web server, which should ONLY be used for development.