Add ISTS15 website to the Flask app
For some reason, this repository contains two entirely separate
websites.  Don't ask me why.  I moved the single-page website from
earlier and just added it as a route to the Flask app.  The static
content is totally broken, but that's ok.

It's looking like we're gonna have to start all over anyway.