Add orange and red highlighting, plus archive page
The archive page was added so I can actually navigate to the old website
if I want to, because with the removal of the drop-down menu that is no
longer possible.

The orange highlighting is to spice up the page a little bit by
highlighting RIT, RITSEC, and student-built.  This also helps make the
RITSEC logo not stand out as much.

I may have gone a little crazy on the red highlighting, but hey.  It's
pretty great.